refactored the code

+++ training/pascal.py
@@ -2,7 +2,7 @@ def pascal(p):
     def create_row(n):
         row = [1]
         for k in range(n):
-            row.append(row[k] * (n-k) / (k+1))
+            row.append(int(row[k] * (n-k) / (k+1)))
         return row
     tri = []
     for a in range(p):
